arrange recent chat list completed second

+++ app/src/main/java/com/example/hoseoclub/SendChatActivity.java
@@ -2,6 +2,7 @@ package com.example.hoseoclub;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -43,11 +44,12 @@ public class SendChatActivity extends AppCompatActivity {
         sendButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+
                 String content = chatEditText.getText().toString();
                 Object time = ServerValue.TIMESTAMP;
-                MessageItem item = new MessageItem(content, (long) time, 0);
+                MessageItem item = new MessageItem(content, time, 0);
                 senderReference.push().setValue(item);
-                item = new MessageItem(content,(long) time, 1);
+                item = new MessageItem(content,time, 1);
                 receiverReference.push().setValue(item);
                 finish();
             }
